move examples to docs
