+++ website/.vitepress/components/LiveCodes.vue
@@ -7,7 +7,7 @@ import VPSwitch from 'vitepress/dist/client/theme-default/components/VPSwitch.vu
 import LiveCodes from 'livecodes/vue'
 import type { Config, EmbedOptions, Playground } from 'livecodes'
 import { PlaygroundProps } from './PlaygroundProps'
-import { examples } from './examples'
+import { examples } from '../../docs/examples'
 
 const playgroundUrl = 'https://ripple.livecodes.pages.dev'
 const apiUrl = 'https://data.jsdelivr.com/v1/packages/npm/ripple'
